MyData/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
